refactor(presenter): adding app module into the presenter folder
